// src/app/openrqm-api/models/RQMWorkspace.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RQMDocument } from './RQMDocument';
export type RQMWorkspace = {
    id?: number;
    name?: string;
    workspaceId?: number;
    workspaces?: Array<RQMWorkspace>;
    documents?: Array<RQMDocument>;
};